// src/types.ts
export type TokenType =
  | 'comment_start'
  | 'comment_end'
  | 'comment'
  | 'preprocessor'
  | 'defined'
  | 'header_file'
  | 'keyword'
  | 'name'
  | 'string'
  | 'quote double'
  | 'quote single'
  | 'quote angle'
  | 'format_specifier'
  | 'number'
  | 'bin_prefix'
  | 'oct_prefix'
  | 'hex_prefix'
  | 'bin_literal'
  | 'oct_literal'
  | 'hex_literal'
  | 'function'
  | 'operator'
  | 'escape'
  | 'comma'
  | 'semicolon'
  | 'colon'
  | 'hash'
  | 'open_paren'
  | 'close_paren'
  | 'open_curly'
  | 'close_curly'
  | 'open_square'
  | 'close_square'
  | 'whitespace'
  | 'unknown'
  | 'open_paren bracket0'
  | 'open_paren bracket1'
  | 'open_paren bracket2'
  | 'close_paren bracket0'
  | 'close_paren bracket1'
  | 'close_paren bracket2'
  | 'open_curly bracket0'
  | 'open_curly bracket1'
  | 'open_curly bracket2'
  | 'close_curly bracket0'
  | 'close_curly bracket1'
  | 'close_curly bracket2'
  | 'open_square bracket0'
  | 'open_square bracket1'
  | 'open_square bracket2'
  | 'close_square bracket0'
  | 'close_square bracket1'
  | 'close_square bracket2';

export type Token = {
  type: TokenType;
  value: string;
};